requirements.txt
pandas~=2.2.3
requests~=2.32.3
retrying~=1.3.4
python-dotenv~=1.0.1
bs4~=0.0.2
beautifulsoup4~=4.12.3
nltk~=3.9.1
textstat~=0.7.4
transformers~=4.46.2
torch~=2.5.1
torchvision~=0.20.1
torchaudio~=2.5.1
tqdm~=4.67.0
psycopg2